chore(components): rename Icon file
