+++ packages/components/src/components/prompts/ConfirmOnDevice/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled, { css } from 'styled-components';
 import { Image } from '../../Image/Image';
 import { variables, animations } from '../../../config';
-import { Icon } from '../../Icon';
+import { Icon } from '../../Icon/Icon';
 import { DeviceModelInternal } from '@trezor/connect';
 
 enum AnimationDirection {
