+++ packages/components/src/components/typography/Link/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { css } from 'styled-components';
-import { Icon, IconProps } from '../../Icon';
+import { Icon, IconProps } from '../../Icon/Icon';
 import { useTheme } from '../../../utils';
 import { FONT_SIZE } from '../../../config/variables';
 import { ParagraphSize } from '../../../support/types';
